Revert "Implement layout decision heuristic"
This reverts commit 2b80369d94387eaf17b2cdf58bf40ba59b34e659.
